+++ piecrust/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Remove `Into<PathBuf>` bound in `VM::new`
 - Rename `host_debug` export to `hdebug` [#222]
 
 ## [0.5.0] - 2023-06-07